--- study/src/test/java/study/RequestLineTest.java
@@ -1,149 +1,0 @@
-package study;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
-
-import java.util.Arrays;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * HTTP 요청을 파싱하는 RequestLine 구현
- * RequestLine = HTTP 요청의 첫 번째 라인을 의미
- * ex: GET /docs/index.html HTTP/1.1
- */
-public class RequestLineTest {
-    enum HttpMethod {
-        GET, POST;
-
-        private static final Map<String, HttpMethod> NAME_TO_HTTP_METHOD_MAP = Arrays.stream(values())
-                .collect(Collectors.toUnmodifiableMap(Enum::name, httpMethod -> httpMethod));
-
-        public static HttpMethod from(String method) {
-            String upperCase = method.toUpperCase();
-            if (NAME_TO_HTTP_METHOD_MAP.containsKey(upperCase)) {
-                return NAME_TO_HTTP_METHOD_MAP.get(upperCase);
-            }
-
-            throw new IllegalArgumentException("지원되지 않는 HTTP method. 입력=" + upperCase);
-        }
-    }
-
-    static class RequestLine {
-        private static final String REQUEST_LINE_SPLITERATOR = " ";
-        private static final String PATH_QUERY_SPLITERATOR = "&";
-        private static final String QUERY_STRING_KEY_VALUE_SPLITERATOR = "=";
-        private static final String PROTOCOL_VERSION_SPLITERATOR = "/";
-
-        private final HttpMethod method;
-        private final String path;
-        private final String protocol;
-        private final String version;
-        private final Map<String, String> queryStringMap;
-
-        private RequestLine(
-                final HttpMethod method,
-                final String path,
-                final String protocol,
-                final String version,
-                final Map<String, String> queryStringMap
-        ) {
-            this.method = method;
-            this.path = path;
-            this.protocol = protocol;
-            this.version = version;
-            this.queryStringMap = queryStringMap;
-        }
-
-        public static RequestLine from(final String request) {
-            String[] requestLines = request.split("\n");
-            String[] requestLine = requestLines[0].split(REQUEST_LINE_SPLITERATOR);
-
-            String method = requestLine[0];
-
-            String[] pathAndQueryStrings = requestLine[1].split("\\?");
-            String path = pathAndQueryStrings[0];
-            Map<String, String> queryStringMap = createQueryStringMap(pathAndQueryStrings);
-
-            String[] protocolAndVersion = requestLine[2].split(PROTOCOL_VERSION_SPLITERATOR);
-            String protocol = protocolAndVersion[0];
-            String version = protocolAndVersion[1];
-
-            return new RequestLine(HttpMethod.from(method), path, protocol, version, queryStringMap);
-        }
-
-        private static Map<String, String> createQueryStringMap(final String[] pathAndQueryStrings) {
-            String[] queryStrings = pathAndQueryStrings[1].split(PATH_QUERY_SPLITERATOR);
-
-            return Arrays.stream(queryStrings)
-                    .map(queryString -> queryString.split(QUERY_STRING_KEY_VALUE_SPLITERATOR))
-                    .collect(Collectors.toUnmodifiableMap(queryString -> queryString[0], queryString -> queryString[1]));
-        }
-
-        public HttpMethod getMethod() {
-            return method;
-        }
-
-        public String getPath() {
-            return path;
-        }
-
-        public String getProtocol() {
-            return protocol;
-        }
-
-        public String getVersion() {
-            return version;
-        }
-
-        public Map<String, String> getQueryStringMap() {
-            return queryStringMap;
-        }
-    }
-
-    @ParameterizedTest(name = "method = {0}")
-    @EnumSource(value = HttpMethod.class)
-    void 요청에_대한_RequestLine을_파싱한다(HttpMethod method) {
-        final String request = """
-                %s /users HTTP/1.1
-                Host: www.nowhere123.com
-                Accept: image/gif, image/jpeg, */*
-                Accept-Language: en-us
-                Accept-Encoding: gzip, deflate
-                User-Agent: Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)
-                (blank line)
-                """.formatted(method);
-
-        RequestLine requestLine = RequestLine.from(request);
-
-        assertThat(requestLine.getMethod()).isEqualTo(method);
-        assertThat(requestLine.getPath()).isEqualTo("/users");
-        assertThat(requestLine.getProtocol()).isEqualTo("HTTP");
-        assertThat(requestLine.getVersion()).isEqualTo("1.1");
-    }
-
-    @Test
-    void 쿼리_스트링을_파싱한다() {
-        final String request = """
-                GET /users?userId=javajigi&password=password&name=JaeSung HTTP/1.1
-                Host: www.nowhere123.com
-                Accept: image/gif, image/jpeg, */*
-                Accept-Language: en-us
-                Accept-Encoding: gzip, deflate
-                User-Agent: Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)
-                (blank line)
-                """;
-
-        RequestLine requestLine = RequestLine.from(request);
-
-        Map<String, String> queryStringMap = requestLine.getQueryStringMap();
-        assertThat(queryStringMap).containsAllEntriesOf(Map.of(
-                "userId", "javajigi",
-                "password", "password",
-                "name", "JaeSung"
-        ));
-    }
-}
